Tiny refactor
