+++ application/packages/nest-microservices/src/microservices.module.ts
@@ -1,8 +1,8 @@
 import {DynamicModule, Module} from '@nestjs/common'
 import {ClientsModule} from '@nestjs/microservices'
 
-import {NATS_CLIENT_KEY} from './common/clients-module-options/constants'
 import {MicroservicesModuleOptionsInterface} from './interfaces/microservices-module-options.interface'
+import {NATS_CLIENT_KEY} from './services/nats/constants'
 import {NatsService} from './services/nats/nats.service'
 
 @Module({})
